move getPresignedUrl into separate file
