+++ apps/crn-frontend/src/network/teams/Compliance.tsx
@@ -89,7 +89,6 @@ const ComplianceList: React.FC<ComplianceListProps> = ({
   const updateManuscript = usePutManuscript();
   const getComplianceDatasetLink = useDownloadFullComplianceDataset();
   const getComplianceDataset = async () => {
-    console.log('in');
     const presignedUrl = await getComplianceDatasetLink();
     window.open(presignedUrl, '_self');
   };
