+++ apps/crn-frontend/src/network/teams/api.ts
@@ -25,9 +25,9 @@ import {
   ResearchOutputResponse,
   TeamPatchRequest,
   TeamResponse,
-  FileAction,
 } from '@asap-hub/model';
 import { isResearchOutputWorkingGroupRequest } from '@asap-hub/validation';
+import { getPresignedUrl } from '../../shared-api/files';
 import { API_BASE_URL } from '../../config';
 import createListApiUrl from '../../CreateListApiUrl';
 
@@ -579,7 +579,6 @@ export const uploadManuscriptFileViaPresignedUrl = async (
 export const downloadFullComplianceDataset = async (
   authorization: string,
 ): Promise<string> => {
-  console.log('in 2');
   const { presignedUrl: downloadUrl } = await getPresignedUrl(
     'ComplianceFullDataset.csv',
     authorization,
@@ -587,49 +586,5 @@ export const downloadFullComplianceDataset = async (
     'download',
   );
 
-  console.log('what?!');
   return downloadUrl;
-};
-
-export const getPresignedUrl = async (
-  filename: string,
-  authorization: string,
-  contentType?: string,
-  action: FileAction = 'upload',
-): Promise<{ presignedUrl: string }> => {
-  console.log('in 3');
-  const resp = await fetch(`${API_BASE_URL}/files/get-url`, {
-    method: 'POST',
-    headers: {
-      authorization,
-      'Content-Type': 'application/json',
-      ...createSentryHeaders(),
-    },
-    body: JSON.stringify({ filename, action, contentType }),
-  });
-
-  let response;
-  try {
-    response = await resp.json();
-  } catch (error) {
-    throw new BackendError(
-      `Failed to parse JSON response when generating presigned URL. Status: ${resp.status}`,
-      {
-        error: 'ParseError',
-        message: 'Failed to parse JSON response',
-        statusCode: resp.status,
-      },
-      resp.status,
-    );
-  }
-
-  if (!resp.ok) {
-    throw new BackendError(
-      `Failed to generate presigned URL. Expected status 200. Received status ${resp.status}: ${resp.statusText}`,
-      response,
-      resp.status,
-    );
-  }
-
-  return response;
 };
